App/backend/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
pydantic
